requirements.txt
fastapi
uvicorn[standard]
streamlit
pydantic
sentence-transformers
weaviate-client
redis
transformers
numpy
python-dotenv
requests
